Revert "Moved grid logic into reusable component"
This reverts commit 07c5e44631070e64693bc705090de7d5dfbd1dec.

reverted the last commit
